v4.0.3
v4.0.3
- Feature: preload images when Accordian mounts (only preview images are preloaded in Projects section)
- Refactor: remove gradient functions from HelperFunctions file and place them in their own file
- Chore: change unnecessary .tsx files in Utils folder to .ts files
