+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "portfolio",
   "private": true,
-  "version": "4.0.2",
+  "version": "4.0.3",
   "homepage": "https://cielestrial.github.io/",
   "type": "module",
   "scripts": {
