+++ src/components/about/Profile.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { SiCodechef, SiLinkedin, SiGithub } from "react-icons/si";
 
-const logo = "/assets/logo-coloured/android-chrome-512x512.png";
+import myLogo from "/assets/logo-coloured/android-chrome-512x512.png";
 
 const Profile = () => {
   const [effect, setEffect] = useState<"fade-in" | "none">("fade-in");
@@ -33,9 +33,9 @@ const Profile = () => {
           }
         >
           <img
-            src={logo}
+            src={myLogo}
             aria-label="My Logo."
-            alt="Logo"
+            alt="My Logo"
             className="w-full"
             draggable="false"
           />
